Comments not needed code, add validators, set up static files, move not needed files into solution directory
