+++ todo-FullStack/TodoApp/main.py
@@ -1,11 +1,15 @@
 from fastapi import FastAPI
-import models
-from database import engine
+# from database import engine
+from starlette.staticfiles import StaticFiles
+
 from routers import auth, todos
+# import models
 
 app = FastAPI()
 
-models.Base.metadata.create_all(bind=engine)
+# models.Base.metadata.create_all(bind=engine)
 
-app.include_router(auth.router)
+app.mount("/static", StaticFiles(directory="static"), name="static")
+
+# app.include_router(auth.router)
 app.include_router(todos.router)
