+++ todo-FullStack/TodoApp/models.py
@@ -1,30 +1,49 @@
-from sqlalchemy import Boolean, Column, Integer, String, ForeignKey
-from sqlalchemy.orm import relationship
-from database import Base
-
-
-class Users(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True)
-    username = Column(String, unique=True, index=True)
-    first_name = Column(String)
-    last_name = Column(String)
-    hashed_password = Column(String)
-    is_active = Column(Boolean, default=True)
-
-    todos = relationship("Todos", back_populates="owner")
-
-
-class Todos(Base):
-    __tablename__ = "todos"
-
-    id = Column(Integer, primary_key=True, index=True)
-    title = Column(String)
-    description = Column(String)
-    priority = Column(Integer)
-    complete = Column(Boolean, default=False)
-    owner_id = Column(Integer, ForeignKey("users.id"))
-
-    owner = relationship("Users", back_populates="todos")
+# from sqlalchemy import Boolean, Column, Integer, String, ForeignKey
+# from sqlalchemy.orm import relationship
+#
+# from database import Base
+#
+#
+# class Users(Base):
+#     """
+#     User model
+#     - email
+#     - username
+#     - first_name
+#     - last_name
+#     - hashed_password
+#     - is_active
+#     """
+#     __tablename__ = "users"
+#
+#     id = Column(Integer, primary_key=True, index=True)
+#     email = Column(String, unique=True, index=True)
+#     username = Column(String, unique=True, index=True)
+#     first_name = Column(String)
+#     last_name = Column(String)
+#     hashed_password = Column(String)
+#     is_active = Column(Boolean, default=True)
+#
+#     todos = relationship("Todos", back_populates="owner")
+#
+#
+# class Todos(Base):
+#     """
+#     Todo model
+#     - title
+#     - description
+#     - priority
+#     - complete
+#     - owner_id
+#     - owner
+#     """
+#     __tablename__ = "todos"
+#
+#     id = Column(Integer, primary_key=True, index=True)
+#     title = Column(String)
+#     description = Column(String)
+#     priority = Column(Integer)
+#     complete = Column(Boolean, default=False)
+#     owner_id = Column(Integer, ForeignKey("users.id"))
+#
+#     owner = relationship("Users", back_populates="todos")
