+++ todo-FullStack/TodoApp/static/js/jquery-slim.js
