--- todo-FullStack/TodoApp/templates/edit-todo.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <!-- Required meta tags -->
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', path='/todo/css/base.css') }}">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', path='/todo/css/bootstrap.css') }}">
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <title>TodoApp</title>
-
-</head>
-<body>
-
-<div>
-    <nav class="navbar navbar-expand-md navbar-dark main-color fixed-top">
-        <a class="navbar-brand" href="#">Todo App</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
-                aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item active">
-                    <a class="nav-link" href="#">Home </a>
-                </li>
-            </ul>
-        </div>
-    </nav>
-</div>
-
-<div class="container">
-    <div class="card">
-        <div class="card-header">
-             Let's edit your todo!
-        </div>
-        <div class="card-body">
-            <form>
-                <div class="form-group">
-                    <label>Title</label>
-                    <input type="text" class="form-control" name="title"
-                           required>
-                </div>
-                <div class="form-group">
-                    <label>Description</label>
-                    <textarea class="form-control"rows="3" name="description"
-                              required></textarea>
-                </div>
-                <div class="form-group">
-                    <label>Priority</label>
-                    <select class="form-control" name="priority">
-                        <option>1</option>
-                        <option>2</option>
-                        <option>3</option>
-                        <option>4</option>
-                        <option>5</option>
-                    </select>
-                </div>
-                <button type="submit" class="btn btn-primary">Add new todo</button>
-            </form>
-        </div>
-    </div>
-</div>
-
-
-
-<script src="{{ url_for('static', path='/todo/js/jquery-slim.js') }}"></script>
-<script src="{{ url_for('static', path='/todo/js/popper.js') }}"></script>
-<script src="{{ url_for('static', path='/todo/js/bootstrap.js') }}"></script>
-</body>
-</html>
