+++ todo-FullStack/TodoApp/validators.py
@@ -1,0 +1,43 @@
+from pydantic import BaseModel, Field
+from typing import Optional
+
+
+class Todo(BaseModel):
+    title: str
+    description: Optional[str]
+    priority: int = Field(gt=0, lt=6, description="Priority must be between 1 and 5")
+    complete: bool
+
+
+class TodoRequest(BaseModel):
+    title: str = Field(min_length=3)
+    description: str = Field(min_length=3, max_length=100)
+    priority: int = Field(gt=0, lt=6)
+    complete: bool
+
+
+class CreateUserRequest(BaseModel):
+    username: str
+    email: str
+    first_name: str
+    last_name: str
+    password: str
+    role: str
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class CreateUser(BaseModel):
+    username: str
+    email: Optional[str]
+    first_name: str
+    last_name: str
+    password: str
+
+
+class UserVerification(BaseModel):
+    password: str
+    new_password: str = Field(min_length=6)
